Chassis: Add dimension properties
This commit adds support for implementing the dimension properties
as defined in the Redfish Chassis schema.

New DBus interface "Decorator.Dimension" is added with below properties:
1. Depth  - This property represents the depth (length) of the chassis
              in millimeters, as specified by the manufacturer.
2. Height - This property represents the height of the chassis
              in millimeters, as specified by the manufacturer.
3. Width  - This property represents the width of the chassis
              in millimeters, as specified by the manufacturer.

On Nvidia's platforms, the source of this information is firmware running
on GPUs or other entities connected to the GPU and to the BMC.
The information itself is served to the BMC using a proprietary protocol.
The BMC has a daemon that implements this protocol and then hosts the
information on D-Bus.

DBus Interfaces gerrit dependency:
https://gerrit.openbmc-project.xyz/c/openbmc/phosphor-dbus-interfaces/+/43433

Redfish validator passed.

Tested:
$ curl -k -H "X-Auth-Token: $token" -H "Content-Type: application/json"
-X GET https://${bmc}/redfish/v1/Chassis/GPU1
{
  "@odata.id": "/redfish/v1/Chassis/GPU1",
  "@odata.type": "#Chassis.v1_14_0.Chassis",
  "ChassisType": "Module",
  "DepthMm": 900.1,
  "HeightMm": 900.12,
  "Id": "GPU1",
  "Name": "GPU1",
  "WidthMm": 900.11,
  "UUID": "7ee8874a-a7f0-11eb-bcbc-0242ac130002"
}

Signed-off-by: Sharad Yadav <sharady@nvidia.com>
Change-Id: I1c76aeec5ed5bff81da7242bf9c54e85f3c3218d