Move the functions to the chassis utils
These functions can be used by other chassis which is
not exactly same as normal chassis.

eg: Erot chassis is not exactly same as chassis
where we don't implement lot of the properties
which we are implementing in the other chassis
Erot chassis is impelementing soem properties
which other chassis doesn't implement.

TestedBy: Ran the Get operation on the chassis
          to get the chassis data.
          TODO
          Redfish service validator

Signed-off-by: Ratan Gupta <ratankgupta31@gmail.com>