fw-update: Parallel firmware update handling with Activation states.
When one firmware update is in progress, another update should not be
allowed. Currently check is in place while the image is being uploaded
and activated.

However Activation status is not considered while enabling firmware
update in progress flag. Also image is getting copied irrespective of
firmware update status.

This patch adds checks to consider Activation status and disallow file
copy to bmc if firmware update is in progress.

Unit Tests:
  1. 2 parallel update - one update should succeed
     and other should fail with 503
  2. Second Update when firmware activation in progress
     Update should fail with 503
  3. Verify tmp file system. file should not be copied to tmp
     while another update is ongoing
  4. Firmware update once activation is completed. Second update
     should succeed while one update is activated.
  5. Firmware update when prev update is failed. Second update
     should succeed.
  6. Sanity: Sequential updates
  7. Test firmware update with task timeout. Subsequent firmware update
     should be allowed after the timeout.