fw-inventory: Modify software id comparison in fwinventory get.
Currently Id in firmware inventory get is checking the substring. But
there can be a collison if one of the Id is substring of another. For
example ERoT_HMC_Firmware and HMC_Firmware. Due to this inconsistent
behavior was seen.

Modified the Id check to equal comparison to avoid this.

Unit Tests:
  1. Firmware inventory get when one Id is substring of another
  2. Run firmware inventory get in a loop and make sure inventory is
     consistent
  3. Sanity: Firmware inventory get of all other devices