Fixes the behaviour of network protocol implementation
Before this commit it was always assumed that HTTP is not
enabled on the bmcweb, however it is dependent on how the
bmcweb is compiled with, either without SSL support or with
SSL support.

This commit puts the hardcoding behind this compile time
variable.

Tested by:

root@dgx-a100-dp:~# curl http://192.168.31.1/redfish/v1/Managers/bmc/NetworkProtocol
{
  "@odata.id": "/redfish/v1/Managers/bmc/NetworkProtocol",
  "@odata.type": "#ManagerNetworkProtocol.v1_5_0.ManagerNetworkProtocol",
  "Description": "Manager Network Service",
  "FQDN": "hgx",
  "HTTP": {
    "Port": 80,
    "ProtocolEnabled": true
  },
  "HostName": "hgx",
  "IPMI": {
    "Port": null,
    "ProtocolEnabled": false
  },
  "Id": "NetworkProtocol",
  "NTP": {
    "NTPServers": [],
    "ProtocolEnabled": true
  },
  "Name": "Manager Network Protocol",
  "SSH": {
    "Port": 22,
    "ProtocolEnabled": true
  },
  "Status": {
    "Health": "OK",
    "HealthRollup": "OK",
    "State": "Enabled"
  }
 Redfish Validator: Passed

Signed-off-by: Ratan Gupta <ratankgupta31@gmail.com>