spdm: extend timeout from 10s to 20s
Tested in qemu with spdm emulator.
In a real-world scenario where the responder is just slow with gathering measurements this shouldn't hit and the bmcweb timeout should be the limiting factor.
At least assuming the CTExponent is provided with a realistic value.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-5186