fwupdate: Add success message for ForceUpdate PATCH operation
When PATCH operation is successfull on ForceUpdate, empty response is
received, which was incorrect. Also ForceUpdate was set to False when
ForceUpdate was not specified in the request.
This patch adds success message when PATCH operation is successful
on ForceUpdate. Also fixes ForceUpdate setting to False issue when
it's not specified.

Unit Tests:
  1. Set force update to True and verify success message and D-Bus
     object, verify ForceUpdate is set to True in UpdateService
  2. Set force update to False and verify success message and D-Bus
     object, verify ForceUpdate is set to True in UpdateService
  3. Verify ForceUpdate is not set when it's not specified in the PATCH
     request.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-1665

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>