Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-4593 This commit will do the chassis reset through redfish API.
The powercycle operation dependent on the current state of the host. If the host is on,it will map to
xyz.openbmc_project.State.Host.Transition.Reboot otherwise it will map to xyz.openbmc_project.State.Chassis.Transition.PowerCycle

 To summarize the new redfish to xyz.openbmc_project.State.*mappings
 * PowerCycle:
      If host on : xyz.openbmc_project.State.Host.Transition.Reboot
      If host off: xyz.openbmc_project.State.Chassis.Transition.PowerCycle

Testing:

root@lab-30:/home/lab# curl -X POST --user root:0penBmc123@ -H "Content-Type: application/json" -H "If-Match: *" --insecure
 https://10.19.102.18/redfish/v1/Chassis/Luna_Motherboard/Actions/Chassis.Reset -d '{"ResetType" : "PowerCycle"}'

{
  "@Message.ExtendedInfo": [
    {
      "@odata.type": "#Message.v1_1_1.Message",
      "Message": "The request completed successfully.",
      "MessageArgs": [],
      "MessageId": "Base.1.13.0.Success",
      "MessageSeverity": "OK",
      "Resolution": "None"
    }
  ]
}

chassis power cycle is the reset of the full chassis system, but in case it does not apply for some platforms,
it will be reset of the host.

Signed-off-by:Nibin Chacko(nibinc@ami.com)