Aggregation: Fix header clobbering
We attempt to sanitize the response body of the asyncResp object as
the first step of copying over the satellite response into the
object.  This is clobbering the "OData-Version" from the response
header.

We can completely remove this step since the body will already be
empty.

Tested:
OData-Version is still present in the header when querying a satellite
resource
curl -s -D- localhost/redfish/v1/Chassis/5B247A_test | grep OData
OData-Version: 4.0

Signed-off-by: Carson Labrado <clabrado@google.com>
Change-Id: I82f54ae9b990504a3236b7a1273c794b1ee6f090