GetDebugTokenRequest implementation had an error, which caused a race condition when a subsequent request was launched right after the previous one. The previous (older) request's task's DBus match was still active, while the object already had its destructor called. SPDM updates from the newer request caused the old handler to be also called, which resulted in the destructed object being referenced, causing a segmentation fault.
The original bug was less severe and manifested itself on systems which had one or more SPDM certificates missing. If the last SPDM update was generated for an ERoT with a valid certificate, the routine was executed correctly. When the last update was generated for an ERoT with no valid certificate, the final state update did not take place (as it was not triggered in such case).

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-5857
Fixes nvbug https://nvbugswb.nvidia.com/NvBugs5/SWBug.aspx?bugid=3789432&cmtNo=