Mitigate running task getting deleted when max tasks are reached.
Currently when max tasks are reached the task which gets created first
is removed from the queue. However there can be a scenario where task
might be running and which results in 404 task not found error for the
running task.

This patch adds logic to identify completed task which is old to be
removed first. If all tasks are in running state and the task which was
created first will be removed.

Unit Tests:
  1. Verify running task does not gets deleted when there are tasks
     which are completed.
  2. Verify old runnning task gets removed first when all tasks are in
     running state.
  3. Verify old completed task gets removed first when all tasks are in
     completed state.

Fixes nvbug https://nvbugs/3889147

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>