Set TaskStatus in tasks based on message severity
TaskStatus property value in tasks should be based
on message severity as per the spec. If one of the
message is critical and above then status should
be critical. It should be warning if none of the
message is critical but atleast one is warning.
Else it should be OK.

Currently it's set statically based on the task
state.

Unit Tests:
  1. Verify critical, warning and OK message
     scenario with firmware update.
  2. Verify critical, warning and OK message
     scenario with log collection
  3. Verify critical, warning and OK message
     scenario with debug token
  4. Verify critical, warning and OK message
     scenario with spdm measurments

Fixes nvbug https://nvbugs/3923773

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>