Changed order of prechecks for update service
purpose: reordering of prechecks has been introduced to present the error messages of update
firmware in more friendly way.

Tested:

Test case 1 - Try to update firmware when another updare is in progress.

1. Make sure that on hgx the free space is lower than doubled size of firmware image.
For example: if the image is 100MB, the available space should be lower than 200MB.
2. Perform update firmware.

Below, there is example response:

{
  "@odata.id": "/redfish/v1/TaskService/Tasks/0",
  "@odata.type": "#Task.v1_4_3.Task",
  "Id": "0",
  "TaskState": "Running",
  "TaskStatus": "OK"
}

3. Make sure that the update process is in the progress.
3. Try to upload the same file.
4. The redfish endpoint to upload image shouldn't allow to upload the file.
The error message should inform the user that the another update firmware is in the progress.

Below, there is example response:

{
  "error": {
    "@Message.ExtendedInfo": [
      {
        "@odata.type": "#MessageRegistry.v1_4_1.MessageRegistry",
        "Message": "An update is in progress.",
        "MessageArgs": null,
        "MessageId": "Update.1.0.UpdateInProgress",
        "Resolution": "Another update is in progress. Retry the update operation once it is complete.",
        "Severity": "OK"
      }
    ],
    "code": "Update.1.0.UpdateInProgress",
    "message": "An update is in progress."
  }
}

Important: before the change was applied, the error message with information that there is not enough free space is available was displayed.

Test case 2 - Try to update firmware when the size of image is greater than allowed size

1. Prepare image file with size greater than 200MB.
It can be any file, the check to verify the size of the image is first.
For too big package, the procedure should be terminated and appropriate error message returned.

2. Perform update firmware using this image.
4. The redfish endpoint to upload image shouldn't allow to upload the file.
The error message should inform the user that the size of image is greater than allowed size.

Below, there is example response:

{
  "error": {
    "@Message.ExtendedInfo": [
      {
        "@odata.type": "#Message.v1_1_1.Message",
        "Message": "The supplied payload exceeds the maximum size supported by the service.",
        "MessageArgs": [],
        "MessageId": "Base.1.13.0.PayloadTooLarge",
        "MessageSeverity": "Critical",
        "Resolution": "Firmware package size is greater than allowed size. Make sure package size is less than UpdateService.MaxImageSizeBytes property and retry the firmware update operation."
      }
    ],
    "code": "Base.1.13.0.PayloadTooLarge",
    "message": "The supplied payload exceeds the maximum size supported by the service."
  }
}

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-6399

Signed-off-by: Robert Papros <robert.papros@conclusive.pl>