Add support to get gpu processor port connection to the cpu c2c port
This used to assume that the processor was a GPU and that the gpu name
could be gathered from the port object it is connected to's path. For
CPUs the port object is actually at a separate path and does not have
the CPU in it. Support was added to GpuMgr to populate the associated
processor as well so now we get the associated port's path by getting
the name of the port from the port path association that already existed
and get the processor name from the associated processor association
that was added.

Testing

Ran this with updates for other required packages (gpumgr, bmcweb,
and openbmc) and gathered the ports via redfish.
APIs checked:

https://{{BMCIP}}/redfish/v1/Systems/system/Processors/GPU_SXM_1/Ports
https://{{BMCIP}}/redfish/v1/Systems/system/Processors/GPU_SXM_1/Ports/C2C_0
https://{{BMCIP}}/redfish/v1/Systems/system/Processors/CPU_0/Ports/ProcessorModule_0_SysBus_0_NVLink_0_0

Output of https://{{BMCIP}}/redfish/v1/Systems/system/Processors/GPU_SXM_1/Ports/C2C_0
```
{
    "@odata.id": "/redfish/v1/Systems/system/Processors/GPU_SXM_1/Ports/C2C_0",
    "@odata.type": "#Port.v1_4_0.Port",
    "Id": "C2C_0",
    "Links": {
        "ConnectedPorts": [
            {
                "@odata.id": "/redfish/v1/Systems/system/Processors/CPU_0/Ports/ProcessorModule_0_SysBus_0_NVLink_0_0"
            }
        ]
    },
    "Metrics": {
        "@odata.id": "/redfish/v1/Systems/system/Processors/GPU_SXM_1/Ports/C2C_0/Metrics"
    },
    "Name": "C2C_0 Resource",
    "Oem": {
        "Nvidia": {
            "@odata.type": "#NvidiaPort.v1_0_0.NvidiaPort"
        }
    },
    "PortProtocol": "NVLink",
    "PortType": "DownstreamPort",
    "Status": {
        "Conditions": [],
        "Health": "OK",
        "HealthRollup": "OK",
        "State": "Enabled"
    }
}
```

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-7118

Signed-off-by:  Joey Rainville (jrainville@nvidia.com)