upstream_sync: BF: Fix the Redfish Triggers
The sensor name of the BF platform doesn't have the type of the sensor,
but the code of the bmcweb need sensor type to be included in the sensor
name. When user adding triggers, the correct object path of sensors is
needed, which included sensor name and type.
We add the code on the handler of the Refish Trigger schame to search
the correct object path according to the sensor name. The sensor type is
also included in the object path.

Fixes JIRA https://