Fix issue that FirmwareInventory content cannot be properly shown in browser
Root cause:
The Accept field in the http request header from browser doesn't include
"application/json" and instead prefer to use "text/html" which causes
the parsing response to fail in processCollectionResponse(), as it expects
response should be in JSON format

Fix:
Check if the request header has "application/json" in the Accept field or
not. If not, add "application/json" at the beginning of the "Accept" field.

Tested:
Query /redfish/v1/UpdateService/FirmwareInventory in
1. web browser
2. command line

Fixes nvbug https://nvbugspro.nvidia.com/bug/4434706

Signed-off-by: Harry Sung <harrys@nvidia.com>