Moving eMMC Opi Api To Under Managers Resource
The enhancement moves the eMMC opt-in API under Managers Resource.
The reason for changing the location of this functionality is that it is not only used by the firmware update.

The eMMC is enabled by default, so enabling/disabling the state in the Redfish API is not expected.
In this commit, there is no such functionality, only the functionality to present the status of PersistentStorageSettings.

Unit Tests:

1. Execute a similar Redfish request to the one presented below:

curl 192.168.31.1/redfish/v1/Managers/HGX_BMC_0

2. The response should include a section presenting the status of PersistentStorageSettings.

"Oem": {
    "Nvidia": {
      ...
      "PersistentStorageSettings": {
        "Status": {
          "State": "Enabled"
        }
      }
...
	}
...
}

Fixes nvbug https://nvbugs/4458058

Signed-off-by: Robert Papros <rpapros@nvidia.com>