fix bmcweb crash when the multipart firmware update is performed with wrong json format.
with obersvation, try/catch didn't catch the expception when the wrong
json format is parserd and it crashed bmcweb.

Ths solution is not make the parser throw the exception and return error
code. the error message can be sent once the error is returned.
with this solution, bmcweb won't crash when the wrong json is input.

Tested:
perform the multipart fw update with wrong json format and check if
there is the crash.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4783713